PR comments ammendments
